+++ pasers/__init__.py
@@ -1,0 +1,1 @@
+from .parsers import AmazonParser, DigitecParser, OttosParser, Parser, ParserFactory