+++ CHANGELOG.md
@@ -12,6 +12,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 ### Removed
 
+## [0.4.0] - 2022-12-16
+### Added
+- Discord bot commands to control the prices that are being tracked
+- Persistent storage for the products that are being tracked
+
+### Changed
+- The program only interfaces with discord
+
 ## [0.3.0] - 2022-12-16
 ### Added
 - Flexible notification pattern
@@ -28,7 +36,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Minimal solution, everything very static
 - Writing prices in standard output and on discord
 
-[unreleased]: https://github.com/pdMa2s/Scrappy/compare/0.3.0...HEAD
+[unreleased]: https://github.com/pdMa2s/Scrappy/compare/0.4.0...HEAD
+[0.4.0]: https://github.com/pdMa2s/Scrappy/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/pdMa2s/Scrappy/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/pdMa2s/Scrappy/compare/0.1.0...0.2.0
 [0.1.0]: https://github.com/pdMa2s/Scrappy/releases/tag/0.1.0